Cleaned Up Truncation errors on Math and GetWidth/Height Methods

+++ Engine/Math/MathUtils.h
@@ -2,9 +2,9 @@
 
 namespace math
 {
-    const float Pi = 3.14159265359;
-    const float DoublePi = 6.28318530718;
-    const float HalfPi = 1.57079632679;
+    const float Pi = (float)3.14159265359;
+    const float DoublePi = (float)6.28318530718;
+    const float HalfPi = (float)1.57079632679;
 
     int sqr(int v);
 }
